# openapi/trip-service.yaml
openapi: 3.1.0
info:
  title: Trip Service API
  description: Жизненный цикл поездок, расчёт цены, рейтинги, статистика
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Local

tags:
  - name: trips
    description: Управление поездками
  - name: statistics
    description: Аналитика

security:
  - bearerAuth: []

paths:
  /trips:
    post:
      tags: [trips]
      summary: Создать поездку (атомарно резервирует свободного водителя)
      operationId: createTrip
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: |
            Опциональный ключ идемпотентности. Повторный POST с тем же значением
            (для того же passenger_id) вернёт уже созданную поездку, а не создаст
            новую. Защищает от дублей при ретраях клиента после разрыва связи.
          schema:
            type: string
            minLength: 1
            maxLength: 64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Passenger not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: No available drivers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [trips]
      summary: История поездок
      operationId: listTrips
      parameters:
        - name: passenger_id
          in: query
          schema:
            type: integer
            format: int64
        - name: driver_id
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TripStatus'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TripResponse'

  /trips/{id}:
    get:
      tags: [trips]
      summary: Информация о поездке
      operationId: getTrip
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /trips/{id}/status:
    patch:
      tags: [trips]
      summary: Обновить статус поездки
      operationId: updateTripStatus
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripStatusRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /trips/{id}/rating:
    post:
      tags: [trips]
      summary: Оставить оценку (1-5 звёзд)
      operationId: rateTrip
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Trip not in COMPLETED state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /statistics:
    get:
      tags: [statistics]
      summary: Статистика за день
      operationId: getDailyStatistics
      parameters:
        - name: date
          in: query
          required: false
          description: Дата в формате YYYY-MM-DD (по умолчанию — сегодня)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    TripStatus:
      type: string
      enum: [SEARCHING, ASSIGNED, ACCEPTED, IN_PROGRESS, COMPLETED, CANCELLED]

    TripRequest:
      type: object
      required: [passenger_id, origin, destination, distance_km]
      properties:
        passenger_id:
          type: integer
          format: int64
          minimum: 1
        origin:
          type: string
          minLength: 1
          maxLength: 500
        destination:
          type: string
          minLength: 1
          maxLength: 500
        distance_km:
          type: number
          format: double
          minimum: 0.1
          description: Расстояние поездки в километрах для расчёта цены

    TripStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/TripStatus'

    RatingRequest:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5

    TripResponse:
      type: object
      required: [id, passenger_id, status, origin, destination, distance_km, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        passenger_id:
          type: integer
          format: int64
        driver_id:
          type: integer
          format: int64
          nullable: true
        status:
          $ref: '#/components/schemas/TripStatus'
        origin:
          type: string
        destination:
          type: string
        distance_km:
          type: number
          format: double
        price:
          type: number
          format: double
          nullable: true
        rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    StatisticsResponse:
      type: object
      required: [date, total_trips, completed_trips, cancelled_trips, total_revenue, average_rating]
      properties:
        date:
          type: string
          format: date
        total_trips:
          type: integer
          format: int64
        completed_trips:
          type: integer
          format: int64
        cancelled_trips:
          type: integer
          format: int64
        total_revenue:
          type: number
          format: double
        average_rating:
          type: number
          format: double
          nullable: true

    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: string